requirements.txt
Flask-Admin==1.5.0
Flask-API==0.7.1
flask-jsontools==0.1.1.post0
Flask-MySQL==1.4.0
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.2
MySQL-python==1.2.5
SQLAlchemy==1.1.10
